Refactoring

+++ OpenGL_Base/shader.cpp
@@ -67,7 +67,7 @@ void Shader::setInt(const std::string & name, int value) const
 	glUniform1i(glGetUniformLocation(m_id, name.c_str()), value);
 }
 
-void Shader::setFloat(const std::string & name, int value) const
+void Shader::setFloat(const std::string & name, float value) const
 {
 	glUniform1f(glGetUniformLocation(m_id, name.c_str()), value);
 }
